+++ components/server/src/initialization/database.py
@@ -35,6 +35,7 @@ def init_database() -> Database:  # pragma: no cover-behave
     rename_teams_webhook_notification_destination(database)
     rename_axe_selenium_python_to_axe_core(database)
     remove_notification_frequency(database)
+    remove_random_number_source(database)
     return database
 
 
@@ -119,7 +120,7 @@ def remove_random_number_source(database: Database) -> None:  # pragma: no cover
     for report in current_reports(database):
         changed = False
         for metric in metrics(report):
-            for source_uuid, source in metric.get("sources", {}).items():
+            for source_uuid, source in list(metric.get("sources", {}).items()):
                 if source["type"] == "random":
                     del metric["sources"][source_uuid]
                     changed = True